backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "carpe-vitam-api"
version = "0.1.0"
description = "Medical lab results tracker API"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.31.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.31.0",
    "alembic==1.13.2",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt<4.0.0",
    "pdfplumber==0.11.1",
    "python-multipart==0.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.0",
    "ruff==0.5.5",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
